Edited Learning Widget (#3)
* Created 2048 game

Created the 2048 game using cursor

* Finished implementing the 2048 game

Finished the 2048 game, downloads cookies onto the person's computer.

* Added project card for 2048

There is a new project card for the 2048 game.

* Edited Learning Widget

Implemented changes outlined by Alex and Hunter. I reorganized the UI to make the simulation visible while changing the settings. The connective line between the sun and planet also remains when there's an eccentricity of 0.